Split test cases between controller and flask-restx
