+++ tests/test_mongo_controller.py
@@ -3,10 +3,7 @@ from threading import Thread
 import logging
 import sys
 
-import flask
-import flask_restplus
 import pytest
-from layaberr import ValidationFailed
 
 import layabase
 import layabase.mongo
@@ -33,59 +30,12 @@ def controller():
     return controller
 
 
-@pytest.fixture
-def app(controller: layabase.CRUDController):
-    application = flask.Flask(__name__)
-    application.testing = True
-    api = flask_restplus.Api(application)
-    namespace = api.namespace("Test", path="/")
-
-    controller.namespace(namespace)
-
-    @namespace.route("/test")
-    class TestResource(flask_restplus.Resource):
-        @namespace.expect(controller.query_get_parser)
-        @namespace.marshal_with(controller.get_response_model)
-        def get(self):
-            return []
-
-        @namespace.expect(controller.json_post_model)
-        def post(self):
-            return []
-
-        @namespace.expect(controller.json_put_model)
-        def put(self):
-            return []
-
-        @namespace.expect(controller.query_delete_parser)
-        def delete(self):
-            return []
-
-    @namespace.route("/test/description")
-    class TestDescriptionResource(flask_restplus.Resource):
-        @namespace.marshal_with(controller.get_model_description_response_model)
-        def get(self):
-            return {}
-
-    @namespace.route("/test_parsers")
-    class TestParsersResource(flask_restplus.Resource):
-        @namespace.expect(controller.query_get_parser)
-        def get(self):
-            return controller.query_get_parser.parse_args()
-
-        @namespace.expect(controller.query_delete_parser)
-        def delete(self):
-            return controller.query_delete_parser.parse_args()
-
-    return application
-
-
 def test_get_all_without_data_returns_empty_list(controller: layabase.CRUDController):
     assert controller.get({}) == []
 
 
 def test_get_one_with_validation_failure(controller: layabase.CRUDController):
-    with pytest.raises(ValidationFailed) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.get_one({"mandatory": "failure"})
     assert exception_info.value.errors == {"mandatory": ["Not a valid int."]}
     assert exception_info.value.received_data == {"mandatory": "failure"}
@@ -111,31 +61,21 @@ def test_get_history(controller: layabase.CRUDController):
 
 
 def test_post_with_nothing_is_invalid(controller: layabase.CRUDController):
-    with pytest.raises(ValidationFailed) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.post(None)
     assert exception_info.value.errors == {"": ["No data provided."]}
     assert not exception_info.value.received_data
 
 
 def test_post_list_with_nothing_is_invalid(controller: layabase.CRUDController):
-    with pytest.raises(ValidationFailed) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.post_many(None)
     assert exception_info.value.errors == {"": ["No data provided."]}
     assert exception_info.value.received_data == []
 
 
 def test_post_with_empty_dict_is_invalid(controller: layabase.CRUDController):
-    with pytest.raises(ValidationFailed) as exception_info:
-        controller.post({})
-    assert exception_info.value.errors == {
-        "key": ["Missing data for required field."],
-        "mandatory": ["Missing data for required field."],
-    }
-    assert exception_info.value.received_data == {}
-
-
-def test_post_with_empty_dict_is_invalid_with_schemas(controller, client):
-    with pytest.raises(ValidationFailed) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.post({})
     assert exception_info.value.errors == {
         "key": ["Missing data for required field."],
@@ -145,63 +85,56 @@ def test_post_with_empty_dict_is_invalid_with_schemas(controller, client):
 
 
 def test_post_with_list_is_invalid(controller: layabase.CRUDController):
-    with pytest.raises(ValidationFailed) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.post([""])
     assert exception_info.value.errors == {"": ["Must be a dictionary."]}
     assert exception_info.value.received_data == [""]
 
 
-def test_post_many_with_dict_is_invalid(controller, client):
-    with pytest.raises(ValidationFailed) as exception_info:
+def test_post_many_with_dict_is_invalid(controller):
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.post_many({""})
     assert exception_info.value.errors == {"": ["Must be a list of dictionaries."]}
     assert exception_info.value.received_data == {""}
 
 
-def test_post_many_with_dict_is_invalid_without_schema(controller):
-    with pytest.raises(ValidationFailed) as exception_info:
-        controller.post_many({""})
-    assert exception_info.value.errors == {"": ["Must be a list."]}
-    assert exception_info.value.received_data == {""}
-
-
 def test_put_with_list_is_invalid(controller: layabase.CRUDController):
-    with pytest.raises(ValidationFailed) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.put([""])
     assert exception_info.value.errors == {"": ["Must be a dictionary."]}
     assert exception_info.value.received_data == [""]
 
 
 def test_put_many_with_dict_is_invalid(controller: layabase.CRUDController):
-    with pytest.raises(ValidationFailed) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.put_many({""})
     assert exception_info.value.errors == {"": ["Must be a list."]}
     assert exception_info.value.received_data == {""}
 
 
 def test_put_many_with_empty_list_is_invalid(controller: layabase.CRUDController):
-    with pytest.raises(ValidationFailed) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.put_many([])
     assert exception_info.value.errors == {"": ["No data provided."]}
     assert exception_info.value.received_data == []
 
 
 def test_post_many_with_empty_list_is_invalid(controller: layabase.CRUDController):
-    with pytest.raises(ValidationFailed) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.post_many([])
     assert exception_info.value.errors == {"": ["No data provided."]}
     assert exception_info.value.received_data == []
 
 
 def test_put_with_nothing_is_invalid(controller: layabase.CRUDController):
-    with pytest.raises(ValidationFailed) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.put(None)
     assert exception_info.value.errors == {"": ["No data provided."]}
     assert exception_info.value.received_data is None
 
 
 def test_put_with_empty_dict_is_invalid(controller: layabase.CRUDController):
-    with pytest.raises(ValidationFailed) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.put({})
     assert exception_info.value.errors == {"key": ["Missing data for required field."]}
     assert exception_info.value.received_data == {}
@@ -212,7 +145,7 @@ def test_delete_without_nothing_do_not_fail(controller: layabase.CRUDController)
 
 
 def test_post_without_mandatory_field_is_invalid(controller: layabase.CRUDController):
-    with pytest.raises(ValidationFailed) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.post({"key": "my_key"})
     assert exception_info.value.errors == {
         "mandatory": ["Missing data for required field."]
@@ -223,7 +156,7 @@ def test_post_without_mandatory_field_is_invalid(controller: layabase.CRUDContro
 def test_post_many_without_mandatory_field_is_invalid(
     controller: layabase.CRUDController,
 ):
-    with pytest.raises(ValidationFailed) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.post_many([{"key": "my_key"}])
     assert exception_info.value.errors == {
         0: {"mandatory": ["Missing data for required field."]}
@@ -232,14 +165,14 @@ def test_post_many_without_mandatory_field_is_invalid(
 
 
 def test_post_without_key_is_invalid(controller: layabase.CRUDController):
-    with pytest.raises(ValidationFailed) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.post({"mandatory": 1})
     assert exception_info.value.errors == {"key": ["Missing data for required field."]}
     assert exception_info.value.received_data == {"mandatory": 1}
 
 
 def test_post_many_without_key_is_invalid(controller: layabase.CRUDController):
-    with pytest.raises(ValidationFailed) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.post_many([{"mandatory": 1}])
     assert exception_info.value.errors == {
         0: {"key": ["Missing data for required field."]}
@@ -248,7 +181,7 @@ def test_post_many_without_key_is_invalid(controller: layabase.CRUDController):
 
 
 def test_post_with_wrong_type_is_invalid(controller: layabase.CRUDController):
-    with pytest.raises(ValidationFailed) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.post({"key": datetime.date(2007, 12, 5), "mandatory": 1})
     assert exception_info.value.errors == {"key": ["Not a valid str."]}
     assert exception_info.value.received_data == {
@@ -266,7 +199,7 @@ def test_post_int_instead_of_str_is_valid(controller: layabase.CRUDController):
 
 
 def test_post_boolean_instead_of_str_is_invalid(controller: layabase.CRUDController):
-    with pytest.raises(ValidationFailed) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.post({"key": True, "mandatory": 1})
     assert {"key": ["Not a valid str."]} == exception_info.value.errors
     assert {"key": True, "mandatory": 1} == exception_info.value.received_data
@@ -283,7 +216,7 @@ def test_rollback_without_versioning_is_valid(controller: layabase.CRUDControlle
 
 
 def test_post_many_with_wrong_type_is_invalid(controller: layabase.CRUDController):
-    with pytest.raises(ValidationFailed) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.post_many([{"key": datetime.date(2007, 12, 5), "mandatory": 1}])
     assert {0: {"key": ["Not a valid str."]}} == exception_info.value.errors
     assert [
@@ -293,7 +226,7 @@ def test_post_many_with_wrong_type_is_invalid(controller: layabase.CRUDControlle
 
 def test_put_with_wrong_type_is_invalid(controller: layabase.CRUDController):
     controller.post({"key": "value1", "mandatory": 1})
-    with pytest.raises(ValidationFailed) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.put({"key": "value1", "mandatory": "invalid value"})
     assert {"mandatory": ["Not a valid int."]} == exception_info.value.errors
     assert {
@@ -310,7 +243,7 @@ def test_put_with_optional_as_none_is_valid(controller: layabase.CRUDController)
 
 def test_put_with_non_nullable_as_none_is_invalid(controller: layabase.CRUDController):
     controller.post({"key": "value1", "mandatory": 1})
-    with pytest.raises(ValidationFailed) as exception_info:
+    with pytest.raises(layabase.ValidationFailed) as exception_info:
         controller.put({"key": "value1", "mandatory": None})
     assert {
         "mandatory": ["Missing data for required field."]
@@ -521,313 +454,6 @@ def test_delete_without_filter_is_removing_everything(
     controller.post({"key": "my_key2", "mandatory": 2, "optional": "my_value2"})
     assert 2 == controller.delete({})
     assert [] == controller.get({})
-
-
-def test_query_get_parser(client):
-    response = client.get("/test_parsers?key=1&mandatory=2&optional=3&limit=4&offset=5")
-    assert response.json == {
-        "key": ["1"],
-        "limit": 4,
-        "mandatory": [2],
-        "offset": 5,
-        "optional": ["3"],
-    }
-
-
-def test_query_delete_parser(client):
-    response = client.delete("/test_parsers?key=1&mandatory=2&optional=3")
-    assert response.json == {"key": ["1"], "mandatory": [2], "optional": ["3"]}
-
-
-def test_open_api_definition(client):
-    response = client.get("/swagger.json")
-    assert response.json == {
-        "swagger": "2.0",
-        "basePath": "/",
-        "paths": {
-            "/test": {
-                "get": {
-                    "responses": {
-                        "200": {
-                            "description": "Success",
-                            "schema": {
-                                "$ref": "#/definitions/TestCollection_GetResponseModel"
-                            },
-                        }
-                    },
-                    "operationId": "get_test_resource",
-                    "parameters": [
-                        {
-                            "name": "key",
-                            "in": "query",
-                            "type": "array",
-                            "items": {"type": "string"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "mandatory",
-                            "in": "query",
-                            "type": "array",
-                            "items": {"type": "integer"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "optional",
-                            "in": "query",
-                            "type": "array",
-                            "items": {"type": "string"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "limit",
-                            "in": "query",
-                            "type": "integer",
-                            "minimum": 0,
-                            "exclusiveMinimum": True,
-                        },
-                        {
-                            "name": "offset",
-                            "in": "query",
-                            "type": "integer",
-                            "minimum": 0,
-                        },
-                        {
-                            "name": "X-Fields",
-                            "in": "header",
-                            "type": "string",
-                            "format": "mask",
-                            "description": "An optional fields mask",
-                        },
-                    ],
-                    "tags": ["Test"],
-                },
-                "delete": {
-                    "responses": {"200": {"description": "Success"}},
-                    "operationId": "delete_test_resource",
-                    "parameters": [
-                        {
-                            "name": "key",
-                            "in": "query",
-                            "type": "array",
-                            "items": {"type": "string"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "mandatory",
-                            "in": "query",
-                            "type": "array",
-                            "items": {"type": "integer"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "optional",
-                            "in": "query",
-                            "type": "array",
-                            "items": {"type": "string"},
-                            "collectionFormat": "multi",
-                        },
-                    ],
-                    "tags": ["Test"],
-                },
-                "put": {
-                    "responses": {"200": {"description": "Success"}},
-                    "operationId": "put_test_resource",
-                    "parameters": [
-                        {
-                            "name": "payload",
-                            "required": True,
-                            "in": "body",
-                            "schema": {
-                                "$ref": "#/definitions/TestCollection_PutRequestModel"
-                            },
-                        }
-                    ],
-                    "tags": ["Test"],
-                },
-                "post": {
-                    "responses": {"200": {"description": "Success"}},
-                    "operationId": "post_test_resource",
-                    "parameters": [
-                        {
-                            "name": "payload",
-                            "required": True,
-                            "in": "body",
-                            "schema": {
-                                "$ref": "#/definitions/TestCollection_PostRequestModel"
-                            },
-                        }
-                    ],
-                    "tags": ["Test"],
-                },
-            },
-            "/test/description": {
-                "get": {
-                    "responses": {
-                        "200": {
-                            "description": "Success",
-                            "schema": {
-                                "$ref": "#/definitions/TestCollection_GetDescriptionResponseModel"
-                            },
-                        }
-                    },
-                    "operationId": "get_test_description_resource",
-                    "parameters": [
-                        {
-                            "name": "X-Fields",
-                            "in": "header",
-                            "type": "string",
-                            "format": "mask",
-                            "description": "An optional fields mask",
-                        }
-                    ],
-                    "tags": ["Test"],
-                }
-            },
-            "/test_parsers": {
-                "delete": {
-                    "responses": {"200": {"description": "Success"}},
-                    "operationId": "delete_test_parsers_resource",
-                    "parameters": [
-                        {
-                            "name": "key",
-                            "in": "query",
-                            "type": "array",
-                            "items": {"type": "string"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "mandatory",
-                            "in": "query",
-                            "type": "array",
-                            "items": {"type": "integer"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "optional",
-                            "in": "query",
-                            "type": "array",
-                            "items": {"type": "string"},
-                            "collectionFormat": "multi",
-                        },
-                    ],
-                    "tags": ["Test"],
-                },
-                "get": {
-                    "responses": {"200": {"description": "Success"}},
-                    "operationId": "get_test_parsers_resource",
-                    "parameters": [
-                        {
-                            "name": "key",
-                            "in": "query",
-                            "type": "array",
-                            "items": {"type": "string"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "mandatory",
-                            "in": "query",
-                            "type": "array",
-                            "items": {"type": "integer"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "optional",
-                            "in": "query",
-                            "type": "array",
-                            "items": {"type": "string"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "limit",
-                            "in": "query",
-                            "type": "integer",
-                            "minimum": 0,
-                            "exclusiveMinimum": True,
-                        },
-                        {
-                            "name": "offset",
-                            "in": "query",
-                            "type": "integer",
-                            "minimum": 0,
-                        },
-                    ],
-                    "tags": ["Test"],
-                },
-            },
-        },
-        "info": {"title": "API", "version": "1.0"},
-        "produces": ["application/json"],
-        "consumes": ["application/json"],
-        "tags": [{"name": "Test"}],
-        "definitions": {
-            "TestCollection_PutRequestModel": {
-                "properties": {
-                    "key": {
-                        "type": "string",
-                        "readOnly": False,
-                        "example": "sample key",
-                    },
-                    "mandatory": {"type": "integer", "readOnly": False, "example": 1},
-                    "optional": {
-                        "type": "string",
-                        "readOnly": False,
-                        "example": "sample optional",
-                    },
-                },
-                "type": "object",
-            },
-            "TestCollection_PostRequestModel": {
-                "properties": {
-                    "key": {
-                        "type": "string",
-                        "readOnly": False,
-                        "example": "sample key",
-                    },
-                    "mandatory": {"type": "integer", "readOnly": False, "example": 1},
-                    "optional": {
-                        "type": "string",
-                        "readOnly": False,
-                        "example": "sample optional",
-                    },
-                },
-                "type": "object",
-            },
-            "TestCollection_GetResponseModel": {
-                "properties": {
-                    "key": {
-                        "type": "string",
-                        "readOnly": False,
-                        "example": "sample key",
-                    },
-                    "mandatory": {"type": "integer", "readOnly": False, "example": 1},
-                    "optional": {
-                        "type": "string",
-                        "readOnly": False,
-                        "example": "sample optional",
-                    },
-                },
-                "type": "object",
-            },
-            "TestCollection_GetDescriptionResponseModel": {
-                "required": ["collection"],
-                "properties": {
-                    "collection": {
-                        "type": "string",
-                        "description": "Collection name",
-                        "example": "collection",
-                    },
-                    "key": {"type": "string", "example": "column"},
-                    "mandatory": {"type": "string", "example": "column"},
-                    "optional": {"type": "string", "example": "column"},
-                },
-                "type": "object",
-            },
-        },
-        "responses": {
-            "ParseError": {"description": "When a mask can't be parsed"},
-            "MaskError": {"description": "When any error occurs on mask"},
-        },
-    }
 
 
 def test_get_with_limit_2_is_retrieving_subset_of_2_first_elements(
